Move spec-section to ingredients
